fixing missing assignment of vapp id
